+++ packages/ocean-areas/src/index.ts
@@ -1,4 +1,5 @@
 import { BBox, Feature, Geometry } from 'geojson'
+import sourceLocales from './data/locales/source.json'
 
 export interface OceanAreaProperties {
   type: string
@@ -9,5 +10,13 @@ export interface OceanAreaProperties {
 }
 
 export type OceanArea = Feature<Geometry, OceanAreaProperties>
+
+export enum OceanAreaLocale {
+  en = 'en',
+  es = 'es',
+  fr = 'fr',
+}
+
+export type OceanAreaLocaleKey = keyof typeof sourceLocales
 
 export * from './ocean-areas'