+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.28.15
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@4.19.2
+  - @globalfishingwatch/react-hooks@5.3.2
+  - @globalfishingwatch/ui-components@2.17.3
+
+## 0.28.14
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/ocean-areas@0.2.1
+
+## 0.28.13
+
+### Patch Changes
+
+- e284ce5e: Fishing map/improve miniglobe performance
+- Updated dependencies [e284ce5e]
+  - @globalfishingwatch/ui-components@2.17.2
+
+## 0.28.12
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/ocean-areas@0.2.0
+
 ## 0.28.11
 
 ### Patch Changes