+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @globalfishingwatch/react-hooks
 
+## 5.3.2
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@4.19.2
+  - @globalfishingwatch/ui-components@2.17.3
+
 ## 5.3.1
 
 ### Patch Changes